occupancy_gen/requirements.txt
absl-py==0.15.0
accelerate==0.20.3
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.20.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
apptools==5.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attr==0.3.2
attrs==23.2.0
awscli==1.32.109
Babel==2.15.0
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
blinker==1.8.2
blosc2==2.0.0
bminf==2.0.1
boto3==1.34.109
botocore==1.34.109
bracex==2.4
brainpp==2.6.4
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cityscapesScripts==2.2.2
clang==5.0
click==8.1.7
click-default-group==1.2.4
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.8
confluent-kafka==2.4.0
contourpy==1.1.1
cpm-kernels==1.0.11
crcmod==1.7
cryptography==42.0.7
cycler==0.12.1
Cython==3.0.10
dacite==1.8.1
dask==2023.3.1
dataclass_array==1.4.1
DCN==1.0
debugpy==1.8.1
decorator==5.1.1
deepdish==0.3.7
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
descartes==1.1.0
detail==0.2.2
diffusers==0.27.2
dm-tree==0.1.8
docutils==0.16
dotmap==1.3.30
easydict==1.10
einops==0.4.1
einsum==0.3.0
elasticsearch==7.17.9
entrypoints==0.4
envisage==7.0.3
etils==1.3.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
fire==0.6.0
flake8==7.1.0
Flask==3.0.3
flatbuffers==1.12
fonttools==4.51.0
formaldict==1.0.6
fqdn==1.5.1
freetype-py==2.4.0
fsspec==2024.5.0
gast==0.4.0
gdown==5.2.0
gitignore_parser==0.1.11
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
h5py==3.9.0
huggingface-hub==0.23.4
humanfriendly==10.0
hydra-core==1.3.0
idna==3.7
imageio==2.31.4
imageio-ffmpeg==0.5.1
imagesize==1.4.1
immutabledict==2.2.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
influxdb==5.3.2
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_packaging==0.12.3
jupyter_server==2.14.0
jupyter_server_fileid==0.9.2
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kapture==1.1.8
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kmatch==0.5.0
kornia==0.5.0
kubernetes==29.0.0
lark==1.1.9
laspy==1.7.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.41.1
locket==1.0.0
loguru==0.7.2
lxml==4.9.3
lyft-dataset-sdk==0.0.8
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.2
matplotlib-inline==0.1.7
mayavi==4.7.3
mccabe==0.7.0
mdurl==0.1.2
megfile==2.2.2
mistune==3.0.2
mmcv==2.0.1
mmcv-full==1.2.7
mmdet==3.3.0
-e git+https://github.com/open-mmlab/mmdetection3d.git@fe25f7a51d36e3702f961e198894580d83c4387b#egg=mmdet3d
mmengine==0.8.4
mpmath==1.3.0
msgpack==1.0.8
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nibabel==5.2.1
nilearn==0.10.4
nltools==0.4.7
nori2==1.12.1
notebook==6.5.7
notebook_shim==0.2.4
numba==0.58.1
numexpr==2.8.6
numpy==1.23.5
numpy-quaternion==2023.0.3
nuscenes-devkit==1.1.10
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.16.0
open3d==0.13.0
opencv-python-headless==4.10.0.84
OpenEXR==1.3.9
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
partd==1.4.1
pathspec==0.12.1
pclpy==0.12.3
pexpect==4.9.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.16.1
pluggy==1.5.0
ply==3.11
plyfile==1.0.3
point-cloud-utils==0.30.0
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==10.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycodestyle==2.12.0
pycolmap==0.4.0
pycparser==2.22
pyface==8.0.0
pyflakes==3.2.0
pyglet==2.0.15
Pygments==2.18.0
pymap3d==3.1.0
PyNaCl==1.5.0
pyntcloud==0.3.1
pynv==0.3
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-Qt5==5.15.14
PyQt5-sip  
pyquaternion==0.9.9
pyrender==0.1.45
PySocks==1.7.1
pytest==7.4.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-statemachine==2.1.2
pytorch-fid==0.3.0
pytz==2024.1
PyVirtualDisplay==3.0
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
redis==5.0.4
referencing==0.35.1
refile==7.2.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.7.2
s3transfer==0.10.1
safetensors==0.4.3
scikit-image==0.19.3
scikit-learn==1.3.0
scipy==1.9.1
seaborn==0.13.2
Send2Trash==1.8.3
Shapely==1.8.5
sip  
six==1.15.0
sk-video==1.1.10
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tables==3.8.0
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.6.0
tensorflow-addons==0.21.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.11.0
tensorflow-gpu==2.6.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.13.0
tensorflow-probability==0.19.0
termcolor==1.1.0
terminado==0.18.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.9.7
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli  
tomlkit==0.12.5
toolz==0.12.1
torch==1.8.1
torchvision==0.9.1
tornado==6.4
tos==2.7.0
tqdm==4.66.4
traitlets==5.14.3
traits==6.4.3
traitsui==8.0.0
transformers==4.44.2
trimesh==4.4.9
typeguard==2.13.3
types-python-dateutil==2.9.0.20240316
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
urllib3==1.26.18
utm==0.7.0
visu3d==1.5.1
vtk==9.0.1
waymo-open-dataset-tf-2-6-0==1.4.9
wcmatch==8.5.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wrapt==1.12.1
wxPython==4.2.1
y-py==0.6.2
yapf==0.40.1
ypy-websocket==0.8.4
zipp==3.18.2
